move dict_keys, dict_items, and dict_values to builtins
