+++ stdlib/_collections_abc.pyi
@@ -1,6 +1,6 @@
 import sys
 from abc import abstractmethod
-from types import MappingProxyType
+from builtins import dict_items as _dict_items, dict_keys as _dict_keys, dict_values as _dict_values
 from typing import (  # noqa: Y022,Y038
     AbstractSet as Set,
     AsyncGenerator as AsyncGenerator,
@@ -12,7 +12,6 @@ from typing import (  # noqa: Y022,Y038
     Container as Container,
     Coroutine as Coroutine,
     Generator as Generator,
-    Generic,
     Hashable as Hashable,
     ItemsView as ItemsView,
     Iterable as Iterable,
@@ -27,9 +26,7 @@ from typing import (  # noqa: Y022,Y038
     Reversible as Reversible,
     Sequence as Sequence,
     Sized as Sized,
-    TypeVar,
     ValuesView as ValuesView,
-    final,
     runtime_checkable,
 )
 
@@ -67,35 +64,12 @@ if sys.version_info < (3, 14):
 if sys.version_info >= (3, 12):
     __all__ += ["Buffer"]
 
-_KT_co = TypeVar("_KT_co", covariant=True)  # Key type covariant containers.
-_VT_co = TypeVar("_VT_co", covariant=True)  # Value type covariant containers.
-
-@final
-class dict_keys(KeysView[_KT_co], Generic[_KT_co, _VT_co]):  # undocumented
-    def __eq__(self, value: object, /) -> bool: ...
-    if sys.version_info >= (3, 13):
-        def isdisjoint(self, other: Iterable[_KT_co], /) -> bool: ...
-    if sys.version_info >= (3, 10):
-        @property
-        def mapping(self) -> MappingProxyType[_KT_co, _VT_co]: ...
-
-@final
-class dict_values(ValuesView[_VT_co], Generic[_KT_co, _VT_co]):  # undocumented
-    if sys.version_info >= (3, 10):
-        @property
-        def mapping(self) -> MappingProxyType[_KT_co, _VT_co]: ...
-
-@final
-class dict_items(ItemsView[_KT_co, _VT_co]):  # undocumented
-    def __eq__(self, value: object, /) -> bool: ...
-    if sys.version_info >= (3, 13):
-        def isdisjoint(self, other: Iterable[tuple[_KT_co, _VT_co]], /) -> bool: ...
-    if sys.version_info >= (3, 10):
-        @property
-        def mapping(self) -> MappingProxyType[_KT_co, _VT_co]: ...
-
 if sys.version_info >= (3, 12):
     @runtime_checkable
     class Buffer(Protocol):
         @abstractmethod
         def __buffer__(self, flags: int, /) -> memoryview: ...
+
+dict_keys = _dict_keys  # undocumented
+dict_items = _dict_items  # undocumented
+dict_values = _dict_values  # undocumented
